putting the abstractions in the library
