+++ src/prelude.rs
@@ -1,0 +1,2 @@
+pub use streaming_iterator::StreamingIterator as _;
+pub use crate::set::Set;
